refactor: move useGetUser hook too hooks/ folder
